autotrader/requirements.txt
alpaca-py>=0.20.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2024.1
yfinance>=0.2.40
flask>=3.0.0
